hotfix : store 타입변환 integer -> Long 으로 변환함
테스트 통과하였음, 컨트롤러 인벤토리 오류 수정해야함.

+++ src/main/java/pinup/backend/store/command/service/StoreService.java
@@ -93,7 +93,7 @@ public class StoreService {
     }
 
     // 아이템 등록(관리자 전용)
-    public Store registerItem(Integer adminId, StoreRequestDto dto) {
+    public Store registerItem(Long adminId, StoreRequestDto dto) {
         Admin admin = adminRepository.findById(adminId)
                 .orElseThrow(() -> new IllegalArgumentException("관리자를 찾을 수 없습니다."));
